requirements.txt
discord.py~=2.6.0
psycopg2-binary~=2.9.10
python-dotenv~=1.1.1
table2ascii~=1.1.3